refactor: move entities to models
