+++ juntagrico_billing/models/__init__.py
@@ -1,7 +1,7 @@
 from django.db.models import signals
 
-from juntagrico_billing.entity.bill import BillItem
-from juntagrico_billing.entity.payment import Payment
+from .bill import BillItem
+from .payment import Payment
 from juntagrico_billing.lifecycle.payment import payment_saved
 from juntagrico_billing.lifecycle.billitem import billitem_saved
